# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WatchStore Api List",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/v1/send_sms": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for send verification sms to user",
                "operationId": "d1247469abd2f7e94eb57030e5a6c480",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "Enter mobile number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/verify_sms_code": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use to check sms code that recieved by user",
                "operationId": "14dd66daa7dd22941ebef438ae54b4e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "user mobile number",
                                        "type": "string"
                                    },
                                    "code": {
                                        "description": "recieved user sms code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/home": {
            "get": {
                "tags": [
                    "Home Page"
                ],
                "description": "get home page data",
                "operationId": "b5746c7cd3b92f1a599cabd61c472a2b",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/most_sold_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "aa401b22c9dfbf112000513a00cf9751",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/most_viewed_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "90bea068d6aebe0f2c22d2b6eec5ea7c",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/newest_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "3e21d83835f3a8b6cf257e841a22e6bc",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/cheapest_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "0f7d18e86e5c9ae05bae556e1df33782",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/most_expensive_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "435cebec62c8fd17d2aa556503189823",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_category/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "7c6ee72b7daee72d814b77a8916c98f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_brand/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "2c9fd69167604725bca0c6023040eed0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_details/{id}": {
            "get": {
                "tags": [
                    "Product Details"
                ],
                "description": "get product details data by product id",
                "operationId": "558d732b5e17b330f599fbcca57c0f62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/save_product_comment": {
            "post": {
                "tags": [
                    "Product Details"
                ],
                "description": "save user comment for product",
                "operationId": "468ce2fc8bc611bf833c75f6191568ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "product id",
                                        "type": "integer"
                                    },
                                    "parent_id": {
                                        "description": "product id",
                                        "type": "integer"
                                    },
                                    "body": {
                                        "description": "user comment text",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data saved",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/search_product": {
            "post": {
                "tags": [
                    "Products Page"
                ],
                "description": "search product",
                "operationId": "595bf7c538a327285caa5db42e67b2ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "User Api"
                ],
                "description": "use to signin user with recieved code",
                "operationId": "5bd3b2ccbefc6cd63d6efda1f88da96e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "user image",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "phone": {
                                        "description": "user phone number",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "user name",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "user address",
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "description": "user postal code",
                                        "type": "string"
                                    },
                                    "lat": {
                                        "description": "user location latitude",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "lang": {
                                        "description": "user location longitude",
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data saved",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}